Dockerize the federator Move abis out of the federator folder Move configs to its own folder inside federator Move integrationTest into its own folder
