+++ .dockerignore
@@ -1,0 +1,8 @@
+./bridge
+./docs
+./ui
+./federator/test
+./federator/integrationTest
+./federator/config
+./federator/node_modules
+./federator/package-lock.json
