+++ docs/TestnetAddresses.md
@@ -1,0 +1,30 @@
+
+# Testnet Addresses and Links
+
+## List of Addresses
+### On RSk
+  - "bridge": [0x684a8a976635fb7ad74a0134ace990a6a0fcce84](https://explorer.testnet.rsk.co/address/0x684a8a976635fb7ad74a0134ace990a6a0fcce84)
+  - "federation": [0x36c893a955399cf15a4a2fbef04c0e06d4d9b379](https://explorer.testnet.rsk.co/address/0x36c893a955399cf15a4a2fbef04c0e06d4d9b379)
+  - "allowTokens": [0x952b706a9ab5fd2d3b36205648ed7852676afbe7](https://explorer.testnet.rsk.co/address/0x952b706a9ab5fd2d3b36205648ed7852676afbe7)
+  - "multiSigWallet": [0x88f6b2bc66f4c31a3669b9b1359524abf79cfc4a](https://explorer.testnet.rsk.co/address/0x88f6b2bc66f4c31a3669b9b1359524abf79cfc4a)
+  - "tRIF": [0x19f64674d8a5b4e652319f5e239efd3bc969a1fe](https://explorer.testnet.rsk.co/address/0x19f64674d8a5b4e652319f5e239efd3bc969a1fe)
+### On Kovan
+  - "bridge": ["0x12ed69359919fc775bc2674860e8fe2d2b6a7b5d"](https://kovan.etherscan.io/address/0x12ed69359919fc775bc2674860e8fe2d2b6a7b5d)
+  - "federation": ["0x8c1901c031cdf42a846c0c422a3b5a2c943f4944"](https://kovan.etherscan.io/address/0x8c1901c031cdf42a846c0c422a3b5a2c943f4944)
+  - "allowTokens": [0xe4aa0f414725c9322a1a9d80d469c5e234786653](https://kovan.etherscan.io/address/0xe4aa0f414725c9322a1a9d80d469c5e234786653)
+  - "multiSigWallet": [0x040007b1804ad78a97f541bebed377dcb60e4138](https://kovan.etherscan.io/address/0x040007b1804ad78a97f541bebed377dcb60e4138)
+  - "etRIF": [0x69f6d4d4813f8e2e618dae7572e04b6d5329e207](https://kovan.etherscan.io/address/0x69f6d4d4813f8e2e618dae7572e04b6d5329e207)
+
+## List of Abis
+[abis used](../abis)
+
+## RSK Testnet Explorer faucet and stats
+- https://stats.testnet.rsk.co/
+- https://explorer.testnet.rsk.co/
+- https://faucet.testnet.rsk.co/
+- https://faucet.rifos.org/
+
+## Kovan (Ethereum testnet) Explorer and faucet
+- https://kovan.etherscan.io/
+- https://faucet.kovan.network/
+- https://gitter.im/kovan-testnet/faucet
