+++ federator/integrationTest/sideTokenBytecode.txt
