+++ federator/src/lib/Federator.js
@@ -1,7 +1,7 @@
 const web3 = require('web3');
 const fs = require('fs');
-const abiBridge = require('../abis/Bridge_v0.json');
-const abiFederation = require('../abis/Federation.json');
+const abiBridge = require('../../../abis/Bridge_v0.json');
+const abiFederation = require('../../../abis/Federation.json');
 const TransactionSender = require('./TransactionSender');
 const CustomError = require('./CustomError');
